requirements.txt
fastmcp
openai
python-dotenv
streamlit
pydantic
nest_asyncio
langchain-community
langchain
langchain-openai